+++ src/Three20UI/Sources/TTTab.m
@@ -72,10 +72,10 @@
       [self addSubview:_badge];
     }
     if (_tabItem.badgeNumber <= kMaxBadgeNumber) {
-      _badge.text = [NSString stringWithFormat:@"%d", _tabItem.badgeNumber];
+      _badge.text = [NSString stringWithFormat:@"%ld", (long)_tabItem.badgeNumber];
 
     } else {
-      _badge.text = [NSString stringWithFormat:@"%d+", kMaxBadgeNumber];
+      _badge.text = [NSString stringWithFormat:@"%ld+", (long)kMaxBadgeNumber];
     }
     [_badge sizeToFit];
